+++ commons/src/main/scala/commons.scala
@@ -486,6 +486,17 @@ object domain {
   }
 }
 
+object MorbidHeaders {
+  val Token        = "X-MorbidToken"
+  val ServiceToken = "X-Morbid-Service-Token"
+}
+
+object MorbidCookies {
+  val Auth         = "morbid-auth"
+  val Token        = "morbid-token"
+  val ServiceToken = "morbid-service-token"
+}
+
 object track {
 
   import morbid.types.*